+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### [Unreleased]
 * Caching of unchanged resources with ETag and If-None-Match headers
- 
+* Avoiding mid-air collisions with ETag and If-Match headers
+
 ### [0.5.0] - 2017-10-20
 * Support for HTTP HEAD Method
 * Experimental GraphQL mutations support